Multiple fixes
* Some servers will not send a single full update, but several. Instead
of using just the first update received, copy all the updates to a
buffer PNG and return that one when we have received enough pixels.

* pngjs does not fully implement the Readable stream interface, and that
causes 'data' events to be lost in the Promise handling, and the
resulting file to be corrupted if you don't use gm, which is a big
dependency we have very little use for. Use the Readable().wrap() method
to resolve the issue with the stream, and remove dependency on gm.

  **Breaking change:** This means the options to resize or change format are no longer
available (they were not documented anyway).

* Remove those options from documentation